+++ CHANGELOG.md
@@ -12,12 +12,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.1.11] - 2024-05-28
+
+### Fixed
+
+- Fixed allowing authed admins to have max permission. This allows them to correctly use `.paa`.
+- Fixed Thunderstore deployment (or it will be soon!)
+
+### Added
+
+- All admin commands are logged. This can be changed by setting the `Command log privilege level = 100` in the XPRising.cfg file.
+
+### Changed
+
+- Auto-saving is no longer chatty.
+
+### Removed
+
+- No longer offer the option of human-readable percentage stats. They are now always in the human-readable format. Not that there are any at the moment.
+
 ## [0.1.10] - 2024-05-28
 
 ### Changed
 
 - Removed admin permission requirement for all permissions. Added a new command (`.paa`) as the only admin command to set the privilege level of the current user to the highest value.
-This allows the configuration of permissions to happen solely within the game.
+  This allows the configuration of permissions to happen solely within the game.
 
 ## [0.1.9] - 2024-05-26